#253 Refactor Standard Schedule

+++ src/main/java/io/github/externschool/planner/service/ScheduleServiceImpl.java
@@ -9,13 +9,14 @@ import io.github.externschool.planner.entity.User;
 import io.github.externschool.planner.entity.schedule.ScheduleEvent;
 import io.github.externschool.planner.entity.schedule.ScheduleEventType;
 import io.github.externschool.planner.entity.schedule.ScheduleHoliday;
-import io.github.externschool.planner.exceptions.UserCannotHandleEventException;
+import io.github.externschool.planner.entity.schedule.ScheduleTemplate;
+import io.github.externschool.planner.exceptions.UserCanNotHandleEventException;
 import io.github.externschool.planner.repository.UserRepository;
 import io.github.externschool.planner.repository.schedule.ParticipantRepository;
 import io.github.externschool.planner.repository.schedule.ScheduleEventRepository;
 import io.github.externschool.planner.repository.schedule.ScheduleEventTypeRepository;
 import io.github.externschool.planner.repository.schedule.ScheduleHolidayRepository;
-import org.springframework.beans.BeanUtils;
+import io.github.externschool.planner.repository.schedule.ScheduleTemplateRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,11 +26,11 @@ import java.time.Duration;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.time.Period;
 import java.time.temporal.ChronoUnit;
 import java.time.temporal.TemporalField;
 import java.time.temporal.WeekFields;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -46,31 +47,34 @@ public class ScheduleServiceImpl implements ScheduleService {
     private final ScheduleEventRepository eventRepository;
     private final ScheduleEventTypeRepository eventTypeRepository;
     private final ParticipantRepository participantRepository;
-    private final UserRepository userRepository;
     private final EmailService emailService;
     private final ScheduleHolidayRepository holidayRepository;
+    private final ScheduleTemplateRepository templateRepository;
+    private final UserRepository userRepository;
 
     private final Executor executor = Executors.newSingleThreadExecutor();
 
     @Autowired
     public ScheduleServiceImpl(final ScheduleEventRepository eventRepository,
                                final ScheduleEventTypeRepository eventTypeRepository,
-                               final UserRepository userRepository,
                                final ParticipantRepository participantRepository,
                                final EmailService emailService,
-                               final ScheduleHolidayRepository holidayRepository) {
+                               final ScheduleHolidayRepository holidayRepository,
+                               final ScheduleTemplateRepository templateRepository,
+                               final UserRepository userRepository) {
         this.eventRepository = eventRepository;
         this.eventTypeRepository = eventTypeRepository;
-        this.userRepository = userRepository;
         this.participantRepository = participantRepository;
         this.emailService = emailService;
         this.holidayRepository = holidayRepository;
+        this.templateRepository = templateRepository;
+        this.userRepository = userRepository;
     }
 
     /**
      * Creates and saves a new event
      *
-     * @return ScheduleEvent
+     * @return Schedule Event, created with provided data
      * @deprecated Replaced by {@link #createEventWithDuration(User owner, ScheduleEventDTO eventDTO, int minutes)}
      */
     @Deprecated
@@ -93,6 +97,14 @@ public class ScheduleServiceImpl implements ScheduleService {
         return this.eventRepository.save(newEvent);
     }
 
+    /**
+     * Creates and saves a new Event, saves provided User with this Event's ownership
+     *
+     * @param owner a User who owns this Event
+     * @param eventDTO provided source DTO
+     * @param minutes duration of the Event in minutes
+     * @return Schedule Event
+     */
     @Override
     public ScheduleEvent createEventWithDuration(final User owner, final ScheduleEventDTO eventDTO, final int minutes) {
         ScheduleEventType type = eventTypeRepository.findByName(eventDTO.getEventType());
@@ -105,11 +117,14 @@ public class ScheduleServiceImpl implements ScheduleService {
                 .withEndDateTime(LocalDateTime.of(eventDTO.getDate(),
                         eventDTO.getStartTime().plus(Duration.of(minutes, ChronoUnit.MINUTES))))
                 .withOwner(owner)
-                .withOpenStatus(eventDTO.getOpen())
                 .withType(type)
+                .withOpenStatus(eventDTO.getOpen() != null ? eventDTO.getOpen() : true)
+                .withCancelledStatus(false)
+                .withAccomplishedStatus(false)
                 .build();
-        eventRepository.save(newEvent);
         owner.addOwnEvent(newEvent);
+        saveEvent(newEvent);
+        userRepository.save(owner);
 
         return newEvent;
     }
@@ -163,6 +178,7 @@ public class ScheduleServiceImpl implements ScheduleService {
             event.getParticipants().forEach(this::removeParticipant);
             Optional.ofNullable(event.getOwner()).ifPresent(owner -> {
                 removeOwner(event);
+                event.setType(null);
                 eventRepository.save(event);
             });
 
@@ -231,8 +247,8 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public Optional<Participant> addParticipant(final User user, final ScheduleEvent event)
-            throws UserCannotHandleEventException {
+    public Optional<Participant> addParticipant(final User user, ScheduleEvent event)
+            throws UserCanNotHandleEventException {
         Optional<Participant> optionalParticipant = Optional.empty();
 
         if (user != null && event != null && event.isOpen()) {
@@ -254,21 +270,19 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public void removeParticipant(final Participant participant) {
-        Optional.ofNullable(participant.getEvent()).map(ScheduleEvent::getParticipants).ifPresent(participants -> {
-            participantRepository.findById(participant.getId())
-                    .ifPresent(p -> {
-                        Optional.ofNullable(participant.getUser()).ifPresent(user -> {
-                            user.removeParticipant(participant);
-                            participantRepository.save(participant);
-                        });
-                        Optional.ofNullable(participant.getEvent()).ifPresent(event -> {
-                            event.removeParticipant(participant);
-                            participantRepository.save(participant);
-                        });
-                        participantRepository.delete(participant);
-                    });
-
-        });
+        if (participant != null) {
+            Optional.ofNullable(participant.getUser()).ifPresent(user -> {
+                user.removeParticipant(participant);
+                userRepository.save(user);
+                participantRepository.save(participant);
+            });
+            Optional.ofNullable(participant.getEvent()).ifPresent(event -> {
+                event.removeParticipant(participant);
+                eventRepository.save(event);
+                participantRepository.save(participant);
+            });
+            participantRepository.delete(participant);
+        }
     }
 
     @Override
@@ -286,7 +300,7 @@ public class ScheduleServiceImpl implements ScheduleService {
     public List<LocalDate> getWeekStartingFirstDay(final LocalDate firstDay) {
         ArrayList<LocalDate> week = new ArrayList<>();
         for (int i = 0; i < 7; i++) {
-            week.add(firstDay.plus(Period.of(0, 0, i)));
+            week.add(firstDay.plusDays(i));
         }
 
         return week;
@@ -294,7 +308,7 @@ public class ScheduleServiceImpl implements ScheduleService {
 
     @Override
     public LocalDate getNextWeekFirstDay() {
-        return getCurrentWeekFirstDay().plus(Period.of(0, 0, 7));
+        return getCurrentWeekFirstDay().plusWeeks(1);
     }
 
     private void canUserOwnAnEventForType(final User user, final ScheduleEventType type) {
@@ -305,7 +319,7 @@ public class ScheduleServiceImpl implements ScheduleService {
                 }
             }
         }
-        throw new UserCannotHandleEventException(
+        throw new UserCanNotHandleEventException(
                 String.format("The user %s is not allowed to create or own %s type of event",
                         user != null ? user.getEmail() : "NULL",
                         type != null ? type.getName() : "NULL"));
@@ -334,62 +348,127 @@ public class ScheduleServiceImpl implements ScheduleService {
     }
 
     @Override
-    public List<ScheduleEvent> createCurrentWeekEventsWithStandardSchemaAndOwner(final User owner) {
-        return eventRepository.saveAll(
-                duplicateStandardEventsByOwnerAndFirstDayOfWeek(owner, getCurrentWeekFirstDay()));
+    public ScheduleTemplate createTemplate(final User owner,
+                                           final ScheduleEventDTO eventDTO,
+                                           final DayOfWeek day,
+                                           final int minutes) {
+        ScheduleEventType type = eventTypeRepository.findByName(eventDTO.getEventType());
+        canUserOwnAnEventForType(owner, type);
+
+        ScheduleTemplate template = ScheduleTemplate.builder()
+                .withTitle(eventDTO.getTitle())
+                .withDescription(eventDTO.getDescription())
+                .withDayOfWeek(day)
+                .withStartOfEvent(eventDTO.getStartTime())
+                .withEndOfEvent(eventDTO.getStartTime().plus(Duration.of(minutes, ChronoUnit.MINUTES)))
+                .withCreatedAt(LocalDateTime.now())
+                .withOwner(owner)
+                .withType(type)
+                .build();
+
+        return templateRepository.save(template);
     }
 
     @Override
-    public List<ScheduleEvent> createNextWeekEventsWithStandardSchemaAndOwner(final User owner) {
-        return eventRepository.saveAll(
-                duplicateStandardEventsByOwnerAndFirstDayOfWeek(owner, getNextWeekFirstDay()));
+    public ScheduleTemplate saveTemplate(final ScheduleTemplate template) {
+        return templateRepository.save(template);
     }
 
-    private List<ScheduleEvent> duplicateStandardEventsByOwnerAndFirstDayOfWeek(final User owner,
-                                                                                final LocalDate firstDay) {
+    @Override
+    public Optional<ScheduleTemplate> findTemplateById(final Long id) {
+        return templateRepository.findById(id);
+    }
+
+    @Override
+    public void deleteTemplateById(final Long id) {
+        templateRepository.deleteById(id);
+    }
+
+    @Override
+    public List<ScheduleTemplate> getTemplatesByOwner(final User owner) {
+        return templateRepository.findAllByOwner(owner);
+    }
+
+    @Override
+    public List<ScheduleEvent> createNextWeekEventsForOwner(final User owner) {
+        return eventRepository.saveAll(createEventsForOwnerByFirstDayOfWeek(owner, getNextWeekFirstDay()));
+    }
+
+    /**
+     *  Since Templates have no Event's LocalDateTime startOfEvent, but DayOfWeek dayOfWeek field only,
+     *  FIRST_MONDAY_OF_EPOCH.plusDays(dayOfWeek.getValue() - 1) are used to define startOfEvent date.
+     *
+     * @param owner Template Events owner user
+     * @return List of Schedule Events converted from stored Templates
+     */
+    @Override
+    public List<ScheduleEvent> getDailyTemplateEventsByOwner(final User owner) {
+        List<ScheduleTemplate> templates = templateRepository.findAllByOwner(owner);
+
+        return templates.stream()
+                .map(template -> createEventForDate(
+                        template,
+                        FIRST_MONDAY_OF_EPOCH.plusDays(template.getDayOfWeek().getValue() - 1)))
+                .collect(Collectors.toList());
+    }
+
+    private List<ScheduleEvent> createEventsForOwnerByFirstDayOfWeek(final User owner,
+                                                                     final LocalDate firstDay) {
         List<ScheduleEvent> newEvents = new ArrayList<>();
-        List<ScheduleEvent> standardSchemaEvents =
-                eventRepository.findAllByOwnerAndStartOfEventBetweenOrderByStartOfEvent(
-                        owner,
-                        LocalDateTime.of(FIRST_MONDAY_OF_EPOCH, LocalTime.MIN),
-                        LocalDateTime.of(FIRST_MONDAY_OF_EPOCH.plusDays(6L), LocalTime.MAX));
-        Set<LocalDate> holidayDates = holidayRepository.findAllByHolidayDateBetween(firstDay, firstDay.plusDays(4))
-                .stream()
-                .map(ScheduleHoliday::getHolidayDate)
-                .collect(Collectors.toSet());
-        // Monday to Friday
-        for (long i = 0; i < 5; i++) {
-            LocalDate date = firstDay.plusDays(i);
+        List<ScheduleTemplate> templates = templateRepository.findAllByOwner(owner);
+        Set<LocalDate> holidayDates = holidayRepository.findAllByHolidayDateBetween(
+                firstDay,
+                firstDay.plusDays(DayOfWeek.FRIDAY.getValue() - 1))
+                    .stream()
+                    .map(ScheduleHoliday::getHolidayDate)
+                    .collect(Collectors.toSet());
+        // Monday to Friday -- create usual working days schedule with templates
+        List<DayOfWeek> daysList = Arrays.asList(
+                DayOfWeek.MONDAY,
+                DayOfWeek.TUESDAY,
+                DayOfWeek.WEDNESDAY,
+                DayOfWeek.THURSDAY,
+                DayOfWeek.FRIDAY);
+        for (DayOfWeek dayOfWeek : daysList) {
+            LocalDate date = firstDay.plusDays(dayOfWeek.getValue() - 1);
             if (!holidayDates.contains(date)) {
-                // this is an ordinary working day, not a holiday
-                standardSchemaEvents.stream()
-                        .filter(event -> event.getStartOfEvent().getDayOfWeek().equals(date.getDayOfWeek()))
-                        .map(event -> duplicateEventForDate(event, date))
+                // this is an ordinary working day, not a holiday, so add it
+                templates.stream()
+                        .filter(template -> template.getDayOfWeek().equals(dayOfWeek))
+                        .map(template -> createEventForDate(template, date))
                         .forEach(newEvents::add);
-            } // do not create events for holidays
+            } // else when current day is a holiday, do not create events
         }
-        // Saturday To Sunday
-        List<ScheduleHoliday> substitutionDays =
-                holidayRepository.findAllBySubstitutionDateBetween(firstDay.plusDays(5L), firstDay.plusDays(6L));
+        // Saturday To Sunday -- add events for the days which are working days this week with holidays' templates
+        List<ScheduleHoliday> substitutionDays = holidayRepository.findAllBySubstitutionDateBetween(
+                firstDay.plusDays(DayOfWeek.SATURDAY.getValue() - 1),
+                firstDay.plusDays(DayOfWeek.SUNDAY.getValue() - 1));
         for (ScheduleHoliday day : substitutionDays) {
-            // this holiday is a substitution working day for another day so fill it with events
+            // this holiday is a substitution working day for another day so fill it with events of that holiday
             DayOfWeek holidayDayOfWeek = day.getHolidayDate().getDayOfWeek();
-            standardSchemaEvents.stream()
-                    .filter(event -> event.getStartOfEvent().getDayOfWeek().equals(holidayDayOfWeek))
-                    .map(event -> duplicateEventForDate(event, day.getSubstitutionDate()))
+            templates.stream()
+                    .filter(template -> template.getDayOfWeek().equals(holidayDayOfWeek))
+                    .map(template -> createEventForDate(template, day.getSubstitutionDate()))
                     .forEach(newEvents::add);
         }
 
         return newEvents;
     }
 
-    private ScheduleEvent duplicateEventForDate(final ScheduleEvent event, final LocalDate date) {
-        ScheduleEvent newEvent = new ScheduleEvent();
-        BeanUtils.copyProperties(event, newEvent);
-        newEvent.setStartOfEvent(LocalDateTime.of(date, LocalTime.from(event.getStartOfEvent())));
-        newEvent.setEndOfEvent(LocalDateTime.of(date, LocalTime.from(event.getEndOfEvent())));
-
-        return newEvent;
+    private ScheduleEvent createEventForDate(final ScheduleTemplate template, final LocalDate date) {
+        return ScheduleEvent.builder()
+                .withId(template.getId())
+                .withTitle(template.getTitle())
+                .withDescription(template.getDescription())
+                .withLocation(template.getLocation())
+                .withOwner(template.getOwner())
+                .withType(template.getType())
+                .withStartDateTime(LocalDateTime.of(date, template.getStartOfEvent()))
+                .withEndDateTime(LocalDateTime.of(date, template.getEndOfEvent()))
+                .withOpenStatus(true)
+                .withCancelledStatus(false)
+                .withAccomplishedStatus(false)
+                .build();
     }
 
     private void canUserParticipateInEventForType(final User user, final ScheduleEventType type) {
@@ -400,7 +479,7 @@ public class ScheduleServiceImpl implements ScheduleService {
                 }
             }
         }
-        throw new UserCannotHandleEventException(
+        throw new UserCanNotHandleEventException(
                 String.format("The user %s is not allowed to participate in %s type of event",
                         user != null ? user.getEmail() : "NULL",
                         type != null ? type.getName() : "NULL"));
